+++ src/components/Game.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import styled from "styled-components";
-import Header from "./Header";
+import Header from "./Board/GameHeader";
 import Board from "../containers/Board";
 import Modal from "../containers/Modal";
