+++ src/containers/Board.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { useDispatch, useSelector } from "react-redux";
-import Board from "../components/Board";
+import Board from "../components/Board/Board";
 import { RootState } from "../rootReducer";
 import {
   openCellAction,
