+++ src/containers/Modal.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { RootState } from "../rootReducer";
-import Modal from "../components/Modal";
+import Modal from "../components/Board/Modal";
 import { createBoardAction } from "../reducers/Board";
 import { gameRetryAction } from "../reducers/Game";
