+++ src/containers/SelectLevel.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { useDispatch } from "react-redux";
-import SelectLevel from "../components/SelectLevel";
+import SelectLevel from "../components/Board/SelectLevel";
 import { SelectLevelAction } from "../reducers/Board";
 import { gameRetryAction } from "../reducers/Game";
 import { LevelKey } from "../reducers/Board/types";
